edited the cli.js file to fix bug-
basically- the command function was taking the current working directory as the default- this means that if the
package gets called outside the actual project folder (the whole point of the library), the cli.js file would be
trying to find "generateor_code_root.js" inside the target directory.

now, we could change the "cwd" option inside the shell helper function to be the directory of the file itself-
"__dirname", however this would reassign the current working directory of the child process spawned by the
cli.js shell method to be the file directory of the library itself- not the target directory. this resulted
in the store being populated inside the library/source directory, not the target

how do we solve this

we pass __dirname inside the command passed to generate call, as an absolute file path, and keep the cwd
option inside the spawn method as process.cwd (which it is by default)

this way, the parent can find the geneator script, and the child procesess are still directed to operate
inside the target directory